delete, edit and add functionality now working.

+++ app/src/main/java/testapp/silencertestapp/Locations.java
@@ -93,7 +93,7 @@ public class Locations {
        return db.insert(CONDITIONS_TBL, null, contentValues);
     }
 
-    public int updateConditionById(long itemId, String wifiName, long start_Time, long end_Time){
+    public int updateConditionById(long itemId, String wifiName, int start_Time, int end_Time){
         ContentValues contentValues = new ContentValues();
         contentValues.put(WIFI_NAME_FIELD, wifiName);
         contentValues.put(START_TIME_FIELD, start_Time);
